Move `filterUserOptionsWithArg` from JavaCompiler to BaseCompiler
So that we could reuse it in other compilers.
